+++ docs/source/ttnn/ttnn/ttnn/multigammaln_bw.rst
@@ -1,0 +1,6 @@
+.. _ttnn.multigammaln_bw:
+
+ttnn.multigammaln_bw
+####################
+
+.. autofunction:: ttnn.multigammaln_bw
