+++ tests/tt_eager/python_api_testing/unit_testing/backward_ops/test_backward_mvlgamma.py
@@ -5,6 +5,7 @@
 import torch
 import pytest
 import tt_lib
+import ttnn
 from tests.tt_eager.python_api_testing.unit_testing.backward_ops.utility_funcs import data_gen_with_range, compare_pcc
 
 
@@ -22,7 +23,7 @@ def test_bw_multigammaln(input_shapes, device):
 
     pyt_y = torch.mvlgamma(in_data, 4)
 
-    tt_output_tensor_on_device = tt_lib.tensor.multigammaln_bw(grad_tensor, input_tensor)
+    tt_output_tensor_on_device = ttnn.multigammaln_bw(grad_tensor, input_tensor)
 
     in_data.retain_grad()
