+++ tt_eager/tt_dnn/op_library/backward/backward_ops.cpp
@@ -2032,29 +2032,6 @@ std::vector<Tensor> complex_sub_bw(
 }
 #undef CHECK_FOR_COMPLEX
 
-std::vector<Tensor> _multigammaln_bw(const Tensor& grad, const Tensor& input, const MemoryConfig& output_mem_config) {
-    std::vector<Tensor> grad_tensor;
-    Tensor digamma_result = ttnn::multiply(grad, digamma(input, output_mem_config), std::nullopt, output_mem_config);
-    Tensor digamma_result_2 = ttnn::multiply(
-        grad, digamma(add_unary(-0.5, input, output_mem_config), output_mem_config), std::nullopt, output_mem_config);
-
-    Tensor grad_result = ttnn::add(digamma_result, digamma_result_2, std::nullopt, output_mem_config);
-
-    digamma_result = ttnn::multiply(
-        grad, digamma(add_unary(-1.0, input, output_mem_config), output_mem_config), std::nullopt, output_mem_config);
-    grad_result = ttnn::add(grad_result, digamma_result, std::nullopt, output_mem_config);
-
-    digamma_result = ttnn::multiply(
-        grad, digamma(add_unary(-1.5, input, output_mem_config), output_mem_config), std::nullopt, output_mem_config);
-    grad_result = ttnn::add(grad_result, digamma_result, std::nullopt, output_mem_config);
-
-    grad_tensor.emplace_back(grad_result);
-    return grad_tensor;
-}
-std::vector<Tensor> multigammaln_bw(const Tensor& grad, const Tensor& input, const MemoryConfig& output_mem_config) {
-    return operation::decorate_as_composite(__func__, _multigammaln_bw)(grad, input, output_mem_config);
-}
-
 // Repeat Backward
 std::vector<Tensor> _repeat_bw(
     const Tensor& grad, const Tensor& input, const Shape& shape, const MemoryConfig& output_mem_config) {
